% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSTVC.R
\name{BSTVC}
\alias{BSTVC}
\title{Bayesian Spatiotemporally Varying Coefficients Model}
\usage{
BSTVC(
    formula = Y ~ ST(X1+X2),
    data,
    study_map,
    Time,
    Space,
    response_type = c("continuous", "binary", "count"),
    threads = 6,
    spatial_matrix=NULL,
    ...
    )
}
\arguments{
\item{formula}{The model formula, with the basic format Y ~ ST(X1 + X2 + X3), is used to specify the target variable and explanatory variables in the modeling. In the BSTVC function, all explanatory variables X must be placed within the ST() symbol, indicating that these explanatory variables have both spatial and temporal nonstationarity (currently only this situation is supported).}

\item{data}{A data frame containing all variables used in the model. If the data involves multiple time sections, it must be in spatiotemporal panel data format. The order of spatial units under each time section must be completely consistent with the order of geographical units in the shapefile.}

\item{study_map}{Data in sf format, imported from map data in shapefile (shp) format, which includes unique value fields for each geographical unit as well as geometric attribute information, etc.}

\item{Time}{A string used to specify the temporal field in the data frame. Only the column name of the temporal field (as a string) needs to be provided.}

\item{Space}{A string used to specify the unique value field representing spatial units in the data frame. Ensure that the string parameter passed in matches the column name in the data data frame exactly, and that data and study_map both have a field with this string name.}

\item{response_type}{A string specifying the type of modeling data, currently supporting three application scenarios: "continuous" for continuous Y variables; "binary" for binary Y variables, where 0 indicates non-occurrence and 1 indicates occurrence; "count" for count Y variables.}

\item{threads}{The number of threads, with a default of 6, which can be set manually.}

\item{spatial_matrix}{The spatial weight matrix, with a default of 10 nearest neighbors matrix. Users can input their own constructed spatial weight matrix files and support multiple types of spatial weight matrices such as k-nearest, inverse distance, fixed distance, etc.; if the user does not input, this function will help calculate the spatial weight matrix under the QUEEN rule (assigning a value of 1 to units with adjacency relationships, otherwise assigning a value of 0).}

\item{...}{NULL}
}
\value{
The BSTVC function outputs a total of six parts, with specific descriptions for each output section as follows:
\enumerate{
\item model.evaluation: Overall evaluation results of the Bayesian model, including commonly used metrics such as deviance information criterion(DIC), watanabe–akaike information criterion(WAIC), and logarithmic score(LS).
\item local.prediction: Local spatiotemporal prediction results for the target variable Y, as well as wide (95\%) and narrow (50\%) Bayesian credible intervals for each prediction value, used to express uncertainty.
\item summary.random.effects: Random effects results, including spatial random effects and temporal random effects for each explanatory variable, along with their wide (95\%) and narrow (50\%) Bayesian credible intervals.
\item time.coefficients: Temporal regression coefficients (temporal nonstationarity), in a wide data format, including the temporal regression coefficients of explanatory variables for each time slice and their wide (95\%) and narrow (50\%) Bayesian credible intervals.
\item space.coefficients: Spatial regression coefficients (spatial nonstationarity), in a wide data format, including the spatial regression coefficients of explanatory variables for each map unit and their wide (95\%) and narrow (50\%) Bayesian credible intervals.
\item STVPI: Spatiotemporal variance partitioning index (STVPI) calculation results, which quantify the percentage of spatiotemporal contribution of each explanatory variable. This not only provides a ranking of factor importance in terms of spatiotemporal heterogeneity but also assesses relative importance. For more information on this tool, please refer to the literature \href{https://doi.org/10.1016/j.jclepro.2022.133781}{Wan et al., 2022}.
}
}
\description{
A class of spatiotemporal regression analysis methods based on Bayesian statistical kernel, whose significant advantage lies in utilizing a "full map" single modeling framework to
uniformly fit the spatiotemporal variations of all local regression coefficients, thereby accurately capturing the spatiotemporal heterogeneity impact of explanatory variables on the target variable, that is, revealing spatiotemporal nonstationarity.
}
\note{
For detailed guidance on data preprocessing, inspection, model fitting, result output, and result visualization, please visit \href{#}{GetStart (Chinese)} or \href{#}{GetStart (English)}.
}
\examples{
\donttest{
# For detailed data processing procedures, modeling steps, and precautions,
# please refer to the document at "vignettes/getstart.Rmd".
data(Florida_Map)
data(Florida_NAT)

# Core field inspection, such as the "Time" field, "Space" field
str(Florida_Map)
str(Florida_NAT)

# Check if the order of sample units for each year in the spatiotemporal panel data table
# is consistent with the order of each spatial unit in the map.
data <- data.check(
  data = Florida_NAT,
  study_map = Florida_Map,
  Time = "Year",
  Space = "FIPS"
)

# Explanatory variable standardization
Florida_NAT[c("DNL","UE","FP","BLK","GI")] <- scale(Florida_NAT[c("DNL","UE","FP","BLK","GI")])

model_1 <- BSTVC(HR ~ ST(DNL + UE + FP + BLK + GI),
  data = Florida_NAT,
  study_map = Florida_Map,
  Time = "Year",
  Space = "FIPS",
  response_type = "continuous",
  threads = 6
)

# class of binary
model_2 <- BSTVC(HW ~ ST(DNL + UE + FP + BLK + GI),
  data = Florida_NAT,
  study_map = Florida_Map,
  Time = "Year",
  Space = "FIPS",
  response_type = "binary",
  threads = 6
)

# class of count
model_3 <- BSTVC(HC ~ ST(DNL + UE + FP + BLK + GI),
  data = Florida_NAT,
  study_map = Florida_Map,
  Time = "Year",
  Space = "FIPS",
  response_type = "count",
  threads = 6
)
}


}
\references{
Chao Song, et al., Exploring spatiotemporal nonstationary effects of climate factors on hand, foot, and mouth disease using Bayesian Spatiotemporally Varying Coefficients (STVC) model in Sichuan, China, Sci. Total Environ. 648 (2019) 550–560.

Chao Song, et al., Spatiotemporally Varying Coefficients (STVC) model: a Bayesian local regression to detect spatial and temporal nonstationarity in variables relationships, ANN GIS 26 (3) (2020) 277–291.

Chao Song, et al., (2022). Spatiotemporal disparities in regional public risk perception of COVID-19 using Bayesian Spatiotemporally Varying Coefficients (STVC) series models across Chinese cities. International Journal of Disaster Risk Reduction, 103078.

Wan Qin, Tang Zhangying, Pan Jay, Xie Mingyu, Wang Shaobin, Yin Hao, Li Junmin, Liu Xin, Yang Yang, & Song Chao. (2022). Spatiotemporal heterogeneity in associations of national population ageing with socioeconomic and environmental factors at the global scale. Journal of Cleaner Production, 373, 133781.

Bayesian STVC series models: \url{https://chaosong.blog/bayesian-stvc/}
}
\seealso{
For those interested in INLA and wanting to learn more about Bayesian modeling and inla modeling, you can refer to \code{\link[INLA]{inla}}.
}
\author{
\href{https://chaosong.blog/}{Chao Song}; Xianteng Tang;
}
