% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSVC.R
\name{BSVC}
\alias{BSVC}
\title{Bayesian Spatially Varying Coefficients Model}
\usage{
BSVC(
    formula =Y~S(X1+X2),
    data,
    study_map,
    Space,
    response_type = c("continuous", "binary", "count"),
    threads = 6,
    spatial_matrix=NULL,
    ...
    )
}
\arguments{
\item{formula}{The model formula, with the basic format Y ~ S(X1 + X2 + X3), is used to specify the target variable and explanatory variables in the modeling. In the BSVC function, all explanatory variables X must be placed within the S() symbol, indicating that the explanatory variables X exhibit spatial nonstationarity.}

\item{data}{A data frame containing all variables used in the model. This data frame should only include information that varies in the spatial dimension, i.e., it should represent a time slice of spatiotemporal panel data, such as data for a specific year or specific month. The order of spatial units in each year in the data must be completely consistent with the order of geographical units in the shapefile.}

\item{study_map}{Data in sf format, imported from map data in shapefile (shp) format, which includes unique value fields for each geographical unit as well as geometric attribute information, etc.}

\item{Space}{A string used to specify the unique value field representing spatial units in the data frame. Ensure that the string parameter passed in matches the column name in the data data frame exactly, and that data and study_map both have a field with this string name.}

\item{response_type}{A string specifying the type of modeling data, currently supporting three application scenarios: "continuous" for continuous Y variables; "binary" for binary Y variables, where 0 indicates non-occurrence and 1 indicates occurrence; "count" for count Y variables.}

\item{threads}{The number of threads, with a default of 6, which can be set manually.}

\item{spatial_matrix}{The spatial weight matrix, with a default of a 10-neighbor matrix. Users can input their own constructed spatial weight matrix files, and support multiple types of spatial weight matrices such as k-nearest, inverse distance, fixed distance, etc.; if the user does not input, this function will help calculate the binary spatial weight matrix under the QUEEN rule (assigning a value of 1 to units with adjacency relationships, otherwise assigning a value of 0).}

\item{...}{NULL}
}
\value{
The BSVC function outputs a total of five parts, with specific descriptions for each output section as follows:
\enumerate{
\item model.evaluation: Overall evaluation results of the Bayesian model, including commonly used metrics such as deviance information criterion(DIC), watanabe–akaike information criterion(WAIC), and logarithmic score(LS).
\item local.prediction: Local prediction results for the target variable Y, along with wide (95\%) and narrow (50\%) Bayesian credible intervals for each predicted value, used to express uncertainty.
\item summary.random.effects: Random effects results, including spatial random effects for each explanatory variable, as well as wide (95\%) and narrow (50\%) Bayesian credible intervals.
\item space.coefficients: Spatial regression coefficients (SCs), in a wide data format, including spatial regression coefficients for each map unit and their wide (95\%) and narrow (50\%) Bayesian credible intervals, used to characterize spatial nonstationarity.
\item STVPI: Spatiotemporal variance partitioning index (STVPI) calculation results, which quantify the spatiotemporal contribution percentage of each explanatory variable. This will output the spatial contribution percentage for each explanatory variable X. For more information on this tool, please refer to the literature \href{https://doi.org/10.1016/j.jclepro.2022.133781}{Wan et al., 2022}.
}
}
\description{
The Bayesian Spatially Varying Coefficients (BSVC) model is a spatially simplified version of the BSTVC model,
designed specifically for identifying variable relationships with spatial heterogeneity, also known as spatial nonstationarity.
Its advantage lies in integrating the "full map" single modeling framework of BSTVC, ensuring that the fitted local spatial regression coefficients are directly comparable.
}
\note{
For detailed guidance on data preprocessing, inspection, model fitting, result output, and result visualization, please visit \href{#}{GetStart (Chinese)} or \href{#}{GetStart (English)}.
}
\examples{
\donttest{
# For detailed data processing procedures, modeling steps, and precautions,
# please refer to the document at "vignettes/getstart.Rmd".
data(Florida_Map)
data(Florida_NAT)

# Core field inspection, such as the "Time" field, "Space" field
str(Florida_Map)
str(Florida_NAT)

# Extract a temporal slice from the data to serve as the input data
# for the spatially varying coefficient model.
data <- Florida_NAT[Florida_NAT$Year == 1990, ]

# Check if the order of sample units for each year in the spatiotemporal panel data table
# is consistent with the order of each spatial unit in the map.
newdata <- data.check(
  data = data,
  study_map = Florida_Map,
  Time = "Year",
  Space = "FIPS"
)

# "Explanatory variable standardization"
newdata[c("DNL", "UE", "FP", "BLK", "GI")] <- scale(newdata[c("DNL", "UE", "FP", "BLK", "GI")])

model_1 <- BSVC(HR ~ S(DNL + UE + FP + BLK + GI),
  data = newdata,
  study_map = Florida_Map,
  Space = "FIPS",
  response_type = "continuous"
)

# class of binary
model_2 <- BSVC(HW ~ S(DNL + UE + FP + BLK + GI),
  data = newdata,
  study_map = Florida_Map,
  Space = "FIPS",
  response_type = "binary"
)

# class of count
model_3 <- BSVC(HC ~ S(DNL + UE + FP + BLK + GI),
  data = newdata,
  study_map = Florida_Map,
  Space = "FIPS",
  response_type = "count"
)
}


}
\references{
Chao Song, et al., Exploring spatiotemporal nonstationary effects of climate factors on hand, foot, and mouth disease using Bayesian Spatiotemporally Varying Coefficients (STVC) model in Sichuan, China, Sci. Total Environ. 648 (2019) 550–560.

Chao Song, et al., Spatiotemporally Varying Coefficients (STVC) model: a Bayesian local regression to detect spatial and temporal nonstationarity in variables relationships, ANN GIS 26 (3) (2020) 277–291.

Chao Song, et al., (2022). Spatiotemporal disparities in regional public risk perception of COVID-19 using Bayesian Spatiotemporally Varying Coefficients (STVC) series models across Chinese cities. International Journal of Disaster Risk Reduction, 103078.

Wan Qin, Tang Zhangying, Pan Jay, Xie Mingyu, Wang Shaobin, Yin Hao, Li Junmin, Liu Xin, Yang Yang, & Song Chao. (2022). Spatiotemporal heterogeneity in associations of national population ageing with socioeconomic and environmental factors at the global scale. Journal of Cleaner Production, 373, 133781.

Bayesian STVC series models: https://chaosong.blog/bayesian-STVC/.
}
\seealso{
For those interested in INLA and wanting to learn more about Bayesian modeling and inla modeling, you can refer to \code{\link[INLA]{inla}}.
}
\author{
\href{https://chaosong.blog/}{Chao Song}; Xianteng Tang;
}
