% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data.check}
\alias{data.check}
\title{Data Checking of Modeling}
\usage{
data.check(data,study_map,Time = "",Space = "")
}
\arguments{
\item{data}{A data frame containing all variables used in the model (including a key field representing unique spatial units, and the type of the key field must be consistent in both data and study_map), with n years of data in a spatiotemporal panel format.}

\item{study_map}{Data in sf format, imported from shapefile (shp) map data. It includes unique value fields for each geographical unit as well as geometric attribute information, etc.}

\item{Time}{A string used to specify the temporal field in the data frame. Ensure that the string parameter passed in matches the column name in the data frame exactly.}

\item{Space}{A string used to specify the unique value field representing spatial units in the data frame. Make sure that the string parameter passed in matches the column name in the data frame exactly, and that both data and study_map have a field with this string name.}
}
\value{
If the order of spatial units in the input data for each year is completely consistent with the order of geographical units in the regional map, only a prompt will be output, meaning users can proceed to the next step of modeling based on the input data. If there is an inconsistency, a newly re-matched dataset will be output.
}
\description{
Based on the key field check, whether the order of spatial units in the spatiotemporal panel data is completely consistent with the order of geographical units in the map. If there is at least one year with an inconsistent order, this function helps to re-match the data.
}
\note{
Note that the number of geographical units in the map data (e.g., 30 districts) should be completely consistent with the number of units in the data; the unique value fields for spatial units (Space) in data and study_map must be of the same format, both numeric, character types, etc.
}
\examples{
data(Florida_NAT)
data(Florida_Map)

# Note: pay attention to check the attributes of each field.
str(Florida_NAT)
str(Florida_Map)
# Florida_Map$FIPS <- as.numeric(Florida_Map$FIPS)

newdata <- data.check(Florida_NAT,
  study_map = Florida_Map,
  Time = "Year",
  Space = "FIPS"
)
}
\seealso{
For methods on converting cross-sectional data (data for each year as a column) into spatiotemporal panel data, refer to the \code{\link[reshape2]{melt}} function in the reshape2 package or the \code{\link[tidyr]{pivot_longer}} function in the tidyr package. For precautions during the data checking process, please refer to the getstart guide in the vignettes folder.
}
