% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Florida_Map}
\alias{Florida_Map}
\title{Map of County Districts in Florida, USA.}
\format{
A object of class \code{\link[sf]{st_read}} with 67 rows and 10 variables:
\describe{
\item{NAME}{County name}
\item{STATE_NAME}{State name}
\item{STATE_FIPS}{State fips code (character)}
\item{CNTY_FIPS}{County fips code (character)}
\item{FIPS}{Combined state and county fips code (character)}
\item{STFIPS}{State fips code (numeric)}
\item{COFIPS}{County fips code (numeric)}
\item{FIPSNO}{Fips code as numeric variable (State fips*1000+County fips)}
\item{SOUTH}{Dummy variable for Southern counties (South = 1)}
\item{geometry}{Geometric information for each county}
}
}
\source{
S. Messner, L. Anselin, D. Hawkins, G. Deane, S. Tolnay, R. Baller (2000). An Atlas of the Spatial Patterning of County-Level Homicide, 1960-1990. Pittsburgh, PA,National Consortium on Violence Research (NCOVR).
Data can be downloaded from the \href{https://geodacenter.github.io/data-and-lab/ncovr/}{GeoDa Data and Lab}.
}
\usage{
Florida_Map
}
\description{
The shapefile data for the 67 counties in Florida, United States, which includes not only unique value fields and county names as attribute information for each county, but also geometric information for each county.
}
\references{
Baller, R., L. Anselin, S. Messner, G. Deane and D. Hawkins (2001). Structural covariates of US county homicide rates: incorporating spatial effects. Criminology 39, 561-590.
}
\keyword{datasets}
